% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcticfiR.R
\docType{package}
\name{arcticfiR}
\alias{arcticfiR}
\alias{arcticfiR-package}
\title{arcticfiR Package}
\description{
arcticfiR is a package built for simple analysis of fire return interval data.
}
\details{
Includes two functions:
\code{\link{is.fire_return}} and
\code{\link{fire_return}}
}
\author{
Brent Thorne \email{brent.thorne18@gmail.com}
}
