% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_return.R
\name{fire_return}
\alias{fire_return}
\title{Average Annual Fire Return}
\usage{
fire_return(x)
}
\arguments{
\item{x}{A data frame where is.fire_return_() returns "TRUE"}
}
\description{
Calculate mean values of years between fire occurances
}
\examples{

my.df <- data.frame(c("a","b","c","d"),
c(1,"None","None","None"),
c("None",1,"None","None"),
c(1,"None","None",1),
c(1,"None",1,"None"),
c("None",1,"None",1),
c(1,"None","None","None"))

colnames(my.df) <- c("ID","Year1","Year2","Year3","Year4","Year5","Year6")

is.fire_return(my.df)

fire_return(x=my.df)

}
