% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.fire.return.R
\name{is.fire_return}
\alias{is.fire_return}
\title{Check Fire Return Data}
\usage{
is.fire_return(x)
}
\arguments{
\item{x}{User's data}
}
\description{
Ensure that user data is suitable for use of the fire_return() function.
}
\examples{
my.df <- data.frame(c("a","b","c","d"),
c(1,"None","None","None"),
c("None",1,"None","None"),
c(1,"None","None",1),
c(1,"None",1,"None"),
c("None",1,"None",1),
c(1,"None","None","None"))

colnames(my.df) <- c("ID","Year1","Year2","Year3","Year4","Year5","Year6")

is.fire_return(my.df)

}
